+++ openrag/services/websearch/providers/staan.py
@@ -1,5 +1,5 @@
 import httpx
-from components.websearch.base import BaseWebSearchProvider, WebResult
+from services.websearch.base import BaseWebSearchProvider, WebResult
 from utils.logger import get_logger
 
 logger = get_logger()
