+++ openrag/services/websearch/service.py
@@ -1,5 +1,5 @@
-from components.websearch.base import BaseWebSearchProvider, WebResult
-from components.websearch.content_fetcher import ContentFetcher
+from services.websearch.base import BaseWebSearchProvider, WebResult
+from services.websearch.content_fetcher import ContentFetcher
 from utils.logger import get_logger
 
 logger = get_logger()
